+++ BUILD
@@ -1533,9 +1533,9 @@ grpc_cc_library(
         "//src/core:bitset",
         "//src/core:call_filters",
         "//src/core:call_final_info",
-        "//src/core:call_size_estimator",
         "//src/core:call_spine",
         "//src/core:cancel_callback",
+        "//src/core:channel",
         "//src/core:channel_args",
         "//src/core:channel_args_endpoint_config",
         "//src/core:channel_args_preconditioning",
